+++ asm8c/m8cmch.c
@@ -78,7 +78,7 @@ struct mne *mp;
 
 	clrexpr(&e1);
 	clrexpr(&e2);
-	op = mp->m_valu;
+	op = (int) mp->m_valu;
 	switch (mp->m_type) {
 
 	case S_MATH:
